pyproject.toml
[project]
name = "scalekit-auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "fastmcp>=3.1.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.0.0",
    "scalekit-sdk-python>=2.5.0",
    "uvicorn>=0.41.0",
]
